feat: Adding validation of the HTTP methods in Client class

+++ application/controllers/api/Client.php
@@ -19,20 +19,24 @@ class Client extends CI_Controller {
     public function getClient()
     {
         try {
-            $authorizationHeader = $this->input->request_headers('Authorization');
-            if (!$authorizationHeader) {
-                return $this->sendJson(['response' => 'Token é necessário.'], 404);
+            if ($this->input->method() === 'get') {
+                $authorizationHeader = $this->input->request_headers('Authorization');
+                if (!$authorizationHeader) {
+                    return $this->sendJson(['response' => 'Token é necessário.'], 404);
+                }
+                if (!$this->isValidClientId()) {
+                    return $this->sendJson(['response' => validation_errors()], 400);
+                }
+                $id = $this->input->post('id');
+                $decodedToken = $this->authorization_token->validateToken($authorizationHeader);
+                if (!$decodedToken['status']) {
+                    return $this->sendJson(['response' => $decodedToken], 200);
+                }
+                $this->data = $this->getClientData($id);
+                return $this->sendJson(['response' => $this->data], 200);
+            }else{
+                return $this->sendJson(['response' => 'Utilize o metodo HTTP correto.'], 500);
             }
-            if (!$this->isValidClientId()) {
-                return $this->sendJson(['response' => validation_errors()], 400);
-            }
-            $id = $this->input->post('id');
-            $decodedToken = $this->authorization_token->validateToken($authorizationHeader);
-            if (!$decodedToken['status']) {
-                return $this->sendJson(['response' => $decodedToken], 200);
-            }
-            $this->data = $this->getClientData($id);
-            return $this->sendJson(['response' => $this->data], 200);
         }catch (Exception $e){
             return $this->sendJson(['response' =>
                 'Ocorreu um erro ao recuperar o cliente.'], 500);
@@ -43,24 +47,31 @@ class Client extends CI_Controller {
      public function createClient()
     {
         try {
-            $authorizationHeader = $this->input->request_headers('Authorization');
-            if (!$authorizationHeader) {
-                return $this->sendJson(['response' => 'Token é necessário.'], 404);
+            if ($this->input->method() === 'post') {
+                $authorizationHeader = $this->input->request_headers('Authorization');
+                if (!$authorizationHeader) {
+                    return $this->sendJson(['response' => 'Token é necessário.'], 404);
+                }
+                if (!$this->isValidFormData()) {
+                    return $this->sendJson(['response' => validation_errors()], 400);
+                }
+                $clientToken = $this->authorization_token->validateToken($authorizationHeader);
+                if (!$clientToken['status']) {
+                    return $this->sendJson(['response' => $clientToken], 500);
+                }
+                $input = $this->input->post();
+                $dataClient['name'] = $this->input->post('name') ?? null;
+                $dataClient['email'] = $this->input->post('email') ?? null;
+                $dataClient['phone'] = $this->input->post('phone') ?? null;
+                $clientID = $this->ClientModel->insert($dataClient);
+                if ($clientID) {
+                    $this->createAddress($input, $clientID);
+                    return $this->sendJson(['response' => 'Cliente registrado com sucesso.'], 200);
+                }
+                return $this->sendJson(['response' => 'Falha ao registrar o cliente.'], 500);
+            }else{
+                return $this->sendJson(['response' => 'Utilize o metodo HTTP correto.'], 500);
             }
-            if (!$this->isValidFormData()) {
-                return $this->sendJson(['response' => validation_errors()], 400);
-            }
-            $clientToken = $this->authorization_token->validateToken($authorizationHeader);
-            if (!$clientToken['status']) {
-                return $this->sendJson(['response' => $clientToken], 500);
-            }
-            $input = $this->input->post();
-            $clientID = $this->ClientModel->insert($input);
-            if ($clientID) {
-                $this->createAddress($input, $clientID);
-                return $this->sendJson(['response' => 'Cliente registrado com sucesso.'], 200);
-            }
-            return $this->sendJson(['response' => 'Falha ao registrar o cliente.'], 500);
         }catch (Exception $e){
             return $this->sendJson(['response' =>
                 'Ocorreu um erro ao criar o cliente.'], 500);
@@ -70,18 +81,22 @@ class Client extends CI_Controller {
     public function updateClient()
     {
         try {
-            $headers = $this->input->request_headers();
-            $id = $this->input->post('id');
-            if (!$this->isValidAuthorization($headers)) {
-                return $this->sendJson(['response' => 'Token é necessário.'], 404);
+            if ($this->input->method() === 'put') {
+                $headers = $this->input->request_headers();
+                $id = $this->input->post('id');
+                if (!$this->isValidAuthorization($headers)) {
+                    return $this->sendJson(['response' => 'Token é necessário.'], 404);
+                }
+                $existingClient = $this->getClientById($id);
+                if (!$existingClient) {
+                    return $this->sendJson(['response' => 'Cliente não encontrado.'], 404);
+                }
+                $input = $this->input->post();
+                $this->updateClientAndAddress($id, $input);
+                return $this->sendJson(['response' => 'Cliente atualizado com sucesso.'], 200);
+            }else{
+                return $this->sendJson(['response' => 'Utilize o metodo HTTP correto.'], 500);
             }
-            $existingClient = $this->getClientById($id);
-            if (!$existingClient) {
-                return $this->sendJson(['response' => 'Cliente não encontrado.'], 404);
-            }
-            $input = $this->input->post();
-            $this->updateClientAndAddress($id, $input);
-            return $this->sendJson(['response' => 'Cliente atualizado com sucesso.'], 200);
         }catch(Exception $e){
             return $this->sendJson(['response' =>
                 'Ocorreu um erro ao atualizar o cliente.'], 500);
@@ -91,20 +106,25 @@ class Client extends CI_Controller {
     public function deleteClient()
     {
         try {
-            $headers = $this->input->request_headers();
-            $id = $this->input->post('id');
-            if (!$this->isValidAuthorization($headers)) {
-                return $this->sendJson(['response' => 'Token é necessário.'], 404);
+            if ($this->input->method() === 'delete') {
+                $headers = $this->input->request_headers();
+                $id = $this->input->post('id');
+                if (!$this->isValidAuthorization($headers)) {
+                    return $this->sendJson(['response' => 'Token é necessário.'], 404);
+                }
+                if (!$this->isValidClientId($id)) {
+                    return $this->sendJson(['response' => validation_errors()], 400);
+                }
+                $existingClient = $this->getClientById($id);
+                if (!$existingClient) {
+                    return $this->sendJson(['response' => 'Cliente não encontrado.'], 404);
+                }
+
+                $this->deleteClientAndAddress($id);
+                return $this->sendJson(['response' => 'Cliente excluído com sucesso.'], 200);
+            }else{
+                return $this->sendJson(['response' => 'Utilize o metodo HTTP correto.'], 500);
             }
-            if (!$this->isValidClientId($id)) {
-                return $this->sendJson(['response' => validation_errors()], 400);
-            }
-            $existingClient = $this->getClientById($id);
-            if (!$existingClient) {
-                return $this->sendJson(['response' => 'Cliente não encontrado.'], 404);
-            }
-            $this->deleteClientAndAddress($id);
-            return $this->sendJson(['response' => 'Cliente excluído com sucesso.'], 200);
         }catch (Exception $e){
             return $this->sendJson(['response' =>
                 'Ocorreu um erro ao excluir o cliente.'], 500);
@@ -113,30 +133,32 @@ class Client extends CI_Controller {
 
     private function sendJson($data)
     {
-        $this->output->set_header('Content-Type: application/json; charset=utf-8')->set_output(json_encode($data));
+        return $this->output->set_header('Content-Type: application/json; charset=utf-8')->set_output(json_encode($data));
     }
 
     private function isValidFormData()
     {
         $rules = [
-            ['field' => 'name', 'label' => 'Name', 'rules' => 'required'],
-            ['field' => 'email', 'label' => 'Email', 'rules' => 'required|valid_email'],
-            ['field' => 'phone', 'label' => 'Phone', 'rules' => 'required'],
+            ['field' => 'name', 'label' => 'Name', 'rules' => 'required', 'errors' => ['required' => 'O campo {field} é obrigatório.']],
+            ['field' => 'email', 'label' => 'Email', 'rules' => 'required|valid_email', 'errors' => [
+                'required' => 'O campo {field} é obrigatório.',
+                'valid_email' => 'Por favor, forneça um endereço de e-mail válido.'
+            ]],
+            ['field' => 'phone', 'label' => 'Phone', 'rules' => 'required', 'errors' => ['required' => 'O campo {field} é obrigatório.']],
         ];
-
         $this->form_validation->set_rules($rules);
         return $this->form_validation->run();
     }
 
-    private function createAddress($input, $clientID)
+    private function createAddress(array $input, int $clientID)
     {
         $addressData = [
-            'cep' => $input['cep'],
-            'street' => $input['street'],
-            'complement' => $input['complement'],
-            'neighborhood' => $input['neighborhood'],
-            'city' => $input['city'],
-            'state' => $input['state'],
+            'cep' => $input['cep'] ?? null,
+            'street' => $input['street'] ?? null,
+            'complement' => $input['complement'] ?? null,
+            'neighborhood' => $input['neighborhood'] ?? null,
+            'city' => $input['city'] ?? null,
+            'state' => $input['state'] ?? null,
             'id_client' => $clientID,
         ];
         $this->AddressModel->insert($addressData);
@@ -166,14 +188,19 @@ class Client extends CI_Controller {
 
     private function updateClientAndAddress($id, $input)
     {
-        $this->ClientModel->update($id, $input);
+        $clientData = [
+            'name' => $input['name'] ?? null,
+            'email' => $input['email'] ?? null,
+            'phone' => $input['phone'] ?? null,
+        ];
+        $this->ClientModel->update($id, $clientData);
         $addressData = [
-            'cep' => $input['cep'],
-            'street' => $input['street'],
-            'complement' => $input['complement'],
-            'neighborhood' => $input['neighborhood'],
-            'city' => $input['city'],
-            'state' => $input['state'],
+            'cep' => $input['cep'] ?? null,
+            'street' => $input['street'] ?? null,
+            'complement' => $input['complement'] ?? null,
+            'neighborhood' => $input['neighborhood'] ?? null,
+            'city' => $input['city'] ?? null,
+            'state' => $input['state'] ?? null,
         ];
         $this->AddressModel->update($id, $addressData);
     }
